infrastructure/foundation/requirements.txt
google-cloud-firestore
google-auth 